Moved final processes to components submodule
