+++ assemblerflow/generator/components/distance_estimation.py
@@ -34,8 +34,8 @@ class PatlasMashDist(Process):
             },
             "refFile": {
                 "default": "'/ngstools/data/patlas.msh'",
-                "description": "Specifies the reference file to be provided to "
-                               "mash. It can either be a fasta or a .msh "
+                "description": "Specifies the reference file to be provided "
+                               "to mash. It can either be a fasta or a .msh "
                                "reference sketch generated by mash."
             }
         }
@@ -62,6 +62,7 @@ class PatlasMashDist(Process):
             "runMashDist",
             "mashDistOutputJson"
         ]
+
 
 class PatlasMashScreen(Process):
 
@@ -91,8 +92,8 @@ class PatlasMashScreen(Process):
             },
             "refFile": {
                 "default": "'/ngstools/data/patlas.msh'",
-                "description": "Specifies the reference file to be provided to "
-                               "mash. It can either be a fasta or a .msh "
+                "description": "Specifies the reference file to be provided "
+                               "to mash. It can either be a fasta or a .msh "
                                "reference sketch generated by mash."
             }
         }
