+++ assemblerflow/generator/components/mapping.py
@@ -17,8 +17,8 @@ class PatlasMapping(Process):
             "max_k": {
                 "default": 10949,
                 "description": "Sets the k parameter for bowtie2 allowing to "
-                               "make multiple mappings of the same read against"
-                               "several hits on the query sequence or "
+                               "make multiple mappings of the same read "
+                               "against several hits on the query sequence or "
                                "sequences."
             },
             "trim5": {
@@ -30,26 +30,26 @@ class PatlasMapping(Process):
             },
             "cov_cutoff": {
                 "default": 0.6,
-                "description": "This variable sets a cutoff for the percentage of "
-                               "the query reference sequence that is covered by "
-                               "reads (in absolute lenght)."
+                "description": "This variable sets a cutoff for the percentage"
+                               " of the query reference sequence that is "
+                               "covered by reads (in absolute lenght)."
             },
             "plasmid_length_dict": {
                 "default": "'jsons/*_length.json'",
                 "description": "Stores a dictionary of lengths to be added to "
                                "jsonDumpingMapping process so that it can "
-                               "easily get the size of each plasmid sequence to"
-                               " be queried."
+                               "easily get the size of each plasmid sequence "
+                               "to be queried."
             },
             "refIndex": {
                 "default": "'/ngstools/data/indexes/bowtie2idx/bowtie2.idx'",
-                "description": "Specifies the reference indexes to be provided "
-                               "to bowtie2."
+                "description": "Specifies the reference indexes to be provided"
+                               " to bowtie2."
             },
             "samtoolsIndex": {
                 "default": "'/ngstools/data/indexes/fasta/samtools.fasta.fai'",
-                "description": "Specifies the reference indexes to be provided "
-                               "to samtools."
+                "description": "Specifies the reference indexes to be provided"
+                               " to samtools."
             },
             "lengthJson": {
                 "default": "'/ngstools/data/reads_sample_result_length.json'",
@@ -65,7 +65,8 @@ class PatlasMapping(Process):
             },
             {
                 "params": "samtoolsIndex",
-                "channel": "IN_samtools_indexes = Channel.value(params.samtoolsIndex)"
+                "channel": "IN_samtools_indexes = Channel"
+                           ".value(params.samtoolsIndex)"
             },
             {
                 "params": "lengthJson",
