+++ assemblerflow/generator/engine.py
@@ -18,7 +18,9 @@ try:
     import generator.components.downloads as downloads
     import generator.components.distance_estimation as distest
     import generator.components.mapping as mapping_patlas
+    import generator.components.mlst as mlst
     import generator.components.reads_quality_control as readsqc
+    import generator.components.typing as typing
     import generator.error_handling as eh
     from __init__ import __version__
     from generator import header_skeleton as hs
@@ -32,7 +34,9 @@ except ImportError:
     import assemblerflow.generator.components.downloads as downloads
     import assemblerflow.generator.components.distance_estimation as distest
     import assemblerflow.generator.components.mapping as mapping_patlas
+    import assemblerflow.generator.components.mlst as mlst
     import assemblerflow.generator.components.reads_quality_control as readsqc
+    import assemblerflow.generator.components.typing as typing
     import assemblerflow.generator.process as pc
     import assemblerflow.generator.error_handling as eh
     from assemblerflow import __version__
@@ -44,8 +48,8 @@ except ImportError:
 process_map = {
         "reads_download": downloads.DownloadReads,
         "integrity_coverage": readsqc.IntegrityCoverage,
-        "seq_typing": pc.SeqTyping,
-        "patho_typing": pc.PathoTyping,
+        "seq_typing": typing.SeqTyping,
+        "patho_typing": typing.PathoTyping,
         "check_coverage": readsqc.CheckCoverage,
         "true_coverage": readsqc.TrueCoverage,
         "fastqc": readsqc.FastQC,
@@ -57,10 +61,10 @@ process_map = {
         "process_skesa": ap.ProcessSkesa,
         "assembly_mapping": ap.AssemblyMapping,
         "pilon": ap.Pilon,
-        "mlst": pc.Mlst,
+        "mlst": mlst.Mlst,
         "abricate": annotation.Abricate,
         "prokka": annotation.Prokka,
-        "chewbbaca": pc.Chewbbaca,
+        "chewbbaca": mlst.Chewbbaca,
         "mash_dist": distest.PatlasMashDist,
         "mash_screen": distest.PatlasMashScreen,
         "mapping_patlas": mapping_patlas.PatlasMapping
