+++ assemblerflow/generator/process.py
@@ -746,289 +746,6 @@ class Init(Process):
         }
 
 
-class SeqTyping(Process):
-    """
-
-    """
-
-    def __init__(self, **kwargs):
-
-        super().__init__(**kwargs)
-
-        self.input_type = "fastq"
-        self.output_type = None
-
-        self.status_channels = []
-
-        self.link_start = None
-
-        self.directives = {"seq_typing": {
-            "cpus": 4,
-            "memory": "'4GB'",
-            "container": "ummidock/seq_typing",
-            "version": "0.1.0-1"
-        }}
-
-        self.params = {
-            "referenceFileO": {
-                "default": "null",
-                "description":
-                    "Fasta file containing reference sequences. If more"
-                    "than one file is passed via the 'referenceFileH parameter"
-                    ", a reference sequence for each file will be determined. "
-                    "(default: $params.referenceFileO)"
-            },
-            "referenceFileH": {
-                "default": "null",
-                "description":
-                    "Fasta file containing reference sequences. If more"
-                    "than one file is passed via the 'referenceFileO parameter"
-                    ", a reference sequence for each file will be determined. "
-                    "(default: $params.referenceFileH)"
-            }
-        }
-
-        self.secondary_inputs = [
-            {
-                "params": "referenceFileO",
-                "channel":
-                    "file(params.referenceFileO) ? params.referenceFileO : "
-                    "exit(1, \"'referenceFileO' parameter missing\")\n"
-                    "IN_refO = Channel"
-                    ".fromPath(params.referenceFileO)"
-                    "map{ it -> it.exists() ? it : exit(1, \"referenceFileO"
-                    " file was not found: '${params.referenceFileO}'\")}"
-            },
-            {
-                "params": "referenceFileH",
-                "channel":
-                    "file(params.referenceFileH) ? params.referenceFileH : "
-                    "exit(1, \"'referenceFileH' parameter missing\")\n"
-                    "IN_refH = Channel"
-                    ".fromPath(params.referenceFileH)"
-                    "map{ it -> it.exists() ? it : exit(1, \"referenceFileH"
-                    " file was not found: '${params.referenceFileH}'\")}"
-            }
-        ]
-
-
-class PathoTyping(Process):
-    """
-
-    """
-
-    def __init__(self, **kwargs):
-
-        super().__init__(**kwargs)
-
-        self.input_type = "fastq"
-        self.output_type = None
-
-        self.ignore_type = True
-
-        self.status_channels = []
-
-        self.params = {
-            "species": {
-                "default": "null",
-                "description":
-                    "Species name. Must be the complete species name with"
-                    "genus and species, e.g.: 'Yersinia enterocolitica'. "
-                    "(default: $params.species)"
-            }
-        }
-
-        self.secondary_inputs = [
-            {
-                "params": "species",
-                "channel":
-                    "if ( !params.species){ exit 1, \"'species' parameter "
-                    "missing\" }\n"
-                    "if ( params.species.toString().split(\" \").size() != 2 )"
-                    "{ exit 1, \"'species' parameter must contain two "
-                    "values (e.g.: 'escherichia coli'). Provided value: "
-                    "${params.species}\"}\n"
-                    "IN_pathoSpecies = Channel.value(params.species)"
-            }
-        ]
-
-        self.link_start = None
-        self.link_end.append({"link": "MAIN_raw",
-                              "alias": "SIDE_PathoType_raw"})
-
-        self.directives = {"patho_typing": {
-            "cpus": 4,
-            "memory": "'4GB'",
-            "container": "ummidock/patho_typing",
-            "version": "0.3.0-1"
-        }}
-
-
-class Mlst(Process):
-    """Mlst mapping process template interface
-
-    This process is set with:
-
-        - ``input_type``: assembly
-        - ``output_type``: None
-        - ``ptype``: post_assembly
-
-    It contains one **secondary channel link end**:
-
-        - ``MAIN_assembly`` (alias: ``MAIN_assembly``): Receives the last
-        assembly.
-    """
-
-    def __init__(self, **kwargs):
-
-        super().__init__(**kwargs)
-
-        self.input_type = "fasta"
-        self.output_type = "fasta"
-
-        self.directives = {"mlst": {
-            "container": "ummidock/mlst",
-        }}
-
-        self.params = {
-            "mlstSpecies": {
-                "default": "null",
-                "description":
-                    "Specify the expected species for MLST checking."
-            }
-        }
-
-
-class Chewbbaca(Process):
-    """Prokka mapping process template interface
-
-    This process is set with:
-
-        - ``input_type``: assembly
-        - ``output_type``: None
-        - ``ptype``: post_assembly
-
-    It contains one **secondary channel link end**:
-
-        - ``MAIN_assembly`` (alias: ``MAIN_assembly``): Receives the last
-        assembly.
-    """
-
-    def __init__(self, **kwargs):
-
-        super().__init__(**kwargs)
-
-        self.input_type = "fasta"
-        self.output_type = None
-
-        self.ignore_type = True
-
-        self.link_start = None
-        self.link_end.append({"link": "MAIN_assembly",
-                              "alias": "MAIN_assembly"})
-
-        self.directives = {
-            "chewbbaca": {
-                "cpus": 4,
-                "container": "mickaelsilva/chewbbaca_py3",
-                "version": "latest",
-            },
-            "chewbbaca_batch": {
-                "cpus": 4,
-                "container": "mickaelsilva/chewbbaca_py3",
-                "version": "latest",
-            },
-            "chewbbacaExtractMLST": {
-                "container": "mickaelsilva/chewbbaca_py3",
-                "version": "latest"
-            }
-        }
-
-        self.params = {
-            "chewbbacaQueue": {
-                "default": "null",
-                "description":
-                    "Specifiy a queue/partition for chewbbaca. This option"
-                    " is only used for grid schedulers. (default: "
-                    "$params.chewbbacaQueue)"
-            },
-            "chewbbacaTraining": {
-                "default": "null",
-                "description":
-                    "Specify the full path to the prodigal training file "
-                    "of the corresponding species. (default: "
-                    "$params.chewbbacaTraining)"
-            },
-            "schemaPath": {
-                "default": "null",
-                "description":
-                    "The path to the chewbbaca schema directory. (default: "
-                    "$params.schemaPath)"
-            },
-            "schemaSelectedLoci": {
-                "default": "null",
-                "description":
-                    "The path to the selection of loci in the schema "
-                    "directory to be used. If not specified, all loci in the"
-                    " schema will be used. (default: "
-                    "$params.schemaSelectedLoci)"
-            },
-            "schemaCore": {
-                "default": "null",
-                "description": ""
-            },
-            "chewbbacaJson": {
-                "default": "false",
-                "description":
-                    "If set to True, chewbbaca's allele call output will be "
-                    "set to JSON format. (default: $params.chewbbacaJson)"
-            },
-            "chewbbacaToPhyloviz": {
-                "default": "false",
-                "description":
-                    "If set to True, the ExtractCgMLST module of chewbbaca"
-                    " will be executed after the allele calling (default: "
-                    "$params.chewbbacaToPhyloviz)",
-            },
-            "chewbbacaProfilePercentage": {
-                "default": 0.95,
-                "description":
-                    "Specifies the proportion of samples that must be "
-                    "present in a locus to save the profile. (default: "
-                    "$params.chewbbacaProfilePercentage)"
-            },
-            "chewbbacaBatch": {
-                "default": "false",
-                "description":
-                    "Specifies whther a chewbbaca run will be performed on the"
-                    " complete input batch (all at the same time) or one by "
-                    "one."
-            }
-        }
-
-        self.secondary_inputs = [
-            {
-                "params": "schemaPath",
-                "channel":
-                    "if ( !params.schemaPath ){ exit 1, \"'schemaPath' "
-                    "parameter missing\"}\n"
-                    "if ( params.chewbbacaTraining){"
-                    "if (!file(params.chewbbacaTraining).exists()) {"
-                    "exit 1, \"'chewbbacaTraining' file was not found: "
-                    "'${params.chewbbacaTraining}'\"}}\n"
-                    "if ( params.schemaSelectedLoci){"
-                    "if (!file(params.schemaSelectedLoci).exists()) {"
-                    "exit 1, \"'schemaSelectedLoci' file was not found: "
-                    "'${params.schemaSelectedLoci}'\"}}\n"
-                    "if ( params.schemaCore){"
-                    "if (!file(params.schemaCore).exists()) {"
-                    "exit 1, \"'schemaCore' file was not found: "
-                    "'${params.schemaCore}'\"}}\n"
-                    "IN_schema = Channel.fromPath(params.schemaPath)"
-            }
-        ]
-
-
 class StatusCompiler(Compiler):
     """Status compiler process template interface
